+++ openair1/PHY/NR_UE_TRANSPORT/nr_transport_ue.h
@@ -63,7 +63,7 @@ typedef struct {
   /// LDPC-code outputs (TS 36.212 V15.4.0, Sec 5.3.2 p. 17)
   uint8_t *e;
   /// Rate matching (Interleaving) outputs (TS 36.212 V15.4.0, Sec 5.4.2.2 p. 30)
-  uint8_t *f;
+  uint32_t *f;
   /// Number of code segments
   uint32_t C;
   /// Number of bits in code segments
